Made some config file changes for supporting some new keywords (LogFile, EventLogging).  Also fixed a slight bug that caused a server to not start properly unless the RootHandle was in its meta handle range (as opposed to any server's meta handle range).  Renamed FS_Name in the config file to be Name instead.  Finally, made sure that anything that can appear in the <Defaults> block can be overridden by placing an entry of the same Tag in the server.conf (for server specific overrides).
Showconfig test program is updated to make sure the parsing logic is sane.
The LogFile and EventLogging fields are not being used actively yet by the
server.


SVN revision 1616

+++ src/common/misc/extent-utils.c
@@ -65,8 +65,8 @@ struct llist *PINT_create_extent_list(char *extent_str)
  */
 int PINT_handle_in_extent(PVFS_handle_extent *ext, PVFS_handle handle)
 {
-    return (((int64_t)handle > ext->first-1) &&
-            ((int64_t)handle < ext->last+1));
+    return ((handle > ext->first-1) &&
+            (handle < ext->last+1));
 }
 
 /* PINT_handle_in_extent_list()
@@ -80,8 +80,9 @@ int PINT_handle_in_extent(PVFS_handle_extent *ext, PVFS_handle handle)
  * otherwise.
  *
  */
-int PINT_handle_in_extent_list(struct llist *extent_list,
-                               PVFS_handle handle)
+int PINT_handle_in_extent_list(
+    struct llist *extent_list,
+    PVFS_handle handle)
 {
     int ret = 0;
     struct llist *cur = NULL;
